rename calcs module to calc. rename peripherals module to peripheral.
