+++ software/deimos/src/calc/constant.rs
