+++ software/deimos/src/calc/inverse_affine.rs
