+++ software/deimos/src/calc/sin.rs
