+++ software/deimos/src/calc/tc_ktype.rs
